rename package to wpilib_interface
