+++ wpilib_interface/src/main/java/coppercore/wpilib_interface/DriveTemplate.java
@@ -1,4 +1,4 @@
-package coppercore.wpi_interface;
+package coppercore.wpilib_interface;
 
 import edu.wpi.first.math.kinematics.ChassisSpeeds;
