+++ wpilib_interface/src/main/java/frc/robot/BuildConstants.java
@@ -1,0 +1,19 @@
+package frc.robot;
+
+/**
+ * Automatically generated file containing build version information.
+ */
+public final class BuildConstants {
+  public static final String MAVEN_GROUP = "com.github.team401";
+  public static final String MAVEN_NAME = "wpi_interface";
+  public static final String VERSION = "0.0";
+  public static final int GIT_REVISION = 31;
+  public static final String GIT_SHA = "6665ac076ea8c0effafb5558d88433ff386cf7b6";
+  public static final String GIT_DATE = "2024-10-07 18:03:40 EDT";
+  public static final String GIT_BRANCH = "drive-command";
+  public static final String BUILD_DATE = "2024-11-11 19:38:24 EST";
+  public static final long BUILD_UNIX_TIME = 1731371904305L;
+  public static final int DIRTY = 1;
+
+  private BuildConstants(){}
+}
